Move managers to the cache package - Not entirely accurate name but not really worth to have a separate package for managers
